ll_ocadr/pyproject.toml
[project]
name = "ll-ocadr"
version = "0.1.0"
description = "Lattice Labs Optical CAD Recognition - 3D Geometry Processing for LLMs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Lattice Labs", email = "noreply@latticelabs.com"}
]
keywords = ["3D", "CAD", "STEP", "STL", "OBJ", "mesh", "vision-language", "multimodal", "LLM"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "vllm>=0.2.0",
    "trimesh>=3.23.0",
    "pythonocc-core>=7.7.0",
    "open3d>=0.17.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "huggingface-hub>=0.19.0",
    "tqdm>=4.65.0",
    "pillow>=10.0.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ll_ocadr"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100
